Added navbar, switching branches
