+++ package.json
@@ -8,8 +8,10 @@
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.20.0",
     "bcrypt": "^5.0.0",
+    "bootstrap": "^4.5.2",
     "dotenv": "^8.2.0",
     "react": "^16.13.1",
+    "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3"
